# Tag the module class Dynamic instead of just instances