--- a/torchdynamo/variables/nn_module.py
+++ b/torchdynamo/variables/nn_module.py
@@ -84,7 +84,7 @@ class NNModuleVariable(VariableTracker):
         """Restart analysis treating this module as an UnspecializedNNModuleVariable"""
         mod = tx.output.get_submodule(self.module_key)
         GenerationTracker.tag(mod)
-        # GenerationTracker.mark_class_dynamic(type(mod))
+        GenerationTracker.mark_class_dynamic(type(mod))
         raise RestartAnalysis()
 
     def var_getattr(self, tx, name):