@anijain2305 2022-08-02
Oh, forgot to mention. Jason said that he commented the line earlier there because there could be a setattr in `__init__` function but not in `forward` and we could turn that class as UnSpecialized.

I could guard it on the __init__ name by plumbing the frame name details through the tx.